backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.33.0
dataclasses-json==0.5.9
decorator==5.1.1
dj-database-url==2.0.0
Django==4.2.2
django-addanother==2.2.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.1.0
django-environ==0.10.0
django-extensions==3.1.3
djangorestframework==3.14.0
exceptiongroup==1.1.1
frozenlist==1.3.3
fsspec==2023.6.0
gcsfs==2023.6.0
google-api-core==2.11.1
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.27.0
google-cloud-bigquery==3.11.3
google-cloud-core==2.3.2
google-cloud-resource-manager==1.10.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
langchain==0.0.220
langchainplus-sdk==0.0.19
markdownify==0.11.6
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==23.1
pandas==2.0.2
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.23.3
psycopg2-binary==2.9.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.10
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.10.0
selenium-stealth==1.0.6
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
sqlparse==0.4.4
tenacity==8.2.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
tweepy==4.14.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
wsproto==1.2.0
yarl==1.9.2
